Complete refactoring of code
